requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
aws-lambda-typing==2.20.0
dynaconf==3.2.5
nats-py==2.8.0
